requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
mccabe==0.7.0
pep8-naming==0.13.2
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==1.0.1
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2024.2
